usb: gadget: Relocate fsg_set_toc
The only usage of it is in mass_storage. It works because mass_storage
includes storage_common.c directly. Starting from the next patch the
include is changed from .c to .h, therefore we will have to export this.

So move fsg_set_toc to mass_storage to avoid any unnecessary export.